streamlit/requirements.txt
requests==2.32.3
streamlit==1.34.0
python-dotenv==1.0.0
boto3==1.34.120